+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Fix parallel mode support in Sphinx extension. Closes {issue}`1482`.
+
 ## [7.5.0 (2026-02-02)](https://github.com/kdeldycke/click-extra/compare/v7.4.0...v7.5.0)
 
 - Fix `TableFormat.GITHUB` to render proper alignment hints in the separator row (`:---`, `:---:`, `---:`).